requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docx2txt==0.8
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-pptx==0.6.23
pytz==2024.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.33.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
XlsxWriter==3.2.0
